[build] read uninstall from manifest
This helps us uninstall the right content more generally.
And this also lets us print a warning of the user's versions mismatch
those in the repository (may be intentional, so we just print a warning
so that the user is aware).
This is temporarily a bit noisy, but you can quickly silence it by doing
a simple `make -C deps uninstall` or a more aggressive `rm -rf usr`.